backend/agent_worker/requirements.txt
google-generativeai
python-dotenv>=1.0.1
moviepy>=1.0.3
mcp
google-adk
ffmpeg-python